requirements.txt
openai
fastapi
uvicorn
python-dotenv
requests
apscheduler
transformers
torch
fpdf2
openpyxl